move factory to profile/factory
