+++ internal/cmd/fitgen/profile/factory/factory.tmpl
