+++ internal/cmd/fitgen/profile/mesgdef/mesgdef.tmpl
@@ -8,8 +8,8 @@
 package {{ .Package }}
 
 import (
-	"github.com/muktihari/fit/factory"
 	"github.com/muktihari/fit/internal/sliceutil"
+	"github.com/muktihari/fit/profile/factory"
 	"github.com/muktihari/fit/profile/typedef"
 	"github.com/muktihari/fit/proto"
 	{{- if .Imports -}}
